docker_server_dev/flask_app/requirements.txt
Flask==1.1.2
mysql-connector-python
gunicorn==20.1.0
flask-restplus==0.13.0
# Werkzeug version to avoid issues with flask-restplus
Werkzeug==0.16.1
config
flasgger